serial: 8250_dw: Allow clock reconfiguration for baud change.
Usage of standard termios function for baud change can change the
UART baud, but if soc's needs clock rate change for associated baud
clock then it is not taken care with default termios function.

However the 8250_serial has a wrapper function which takes care
of this issue, so use the wrapper termios function to set the
clock rate before configuring controller for baud change.

Signed-off-by: Govindraj Raja <Govindraj.Raja@imgtec.com>