mtd: nand: Check length of ID before reading bits per cell
The table-based NAND identification currently reads the number
of bits per cell from the 3rd byte of the extended ID. This is done
for the so-called 'full ID' devices; i.e. devices that have a known
length ID.

However, if the ID length is shorter than three, there's no 3rd byte,
and so it's wrong to read the bits per cell from there. Fix this by
adding a check for the ID length.

(picked from http://lists.infradead.org/pipermail/linux-mtd/2014-December/056764.html)

Signed-off-by: Ezequiel Garcia <ezequiel.garcia@imgtec.com>